prepared clevr-tex test

+++ utils/metrics.py
@@ -7,7 +7,7 @@ import torch.nn.functional as F
 
 
 # from scipy.optimize import linear_sum_assignment
-# from sklearn.metrics import adjusted_rand_score
+from sklearn.metrics import adjusted_rand_score
 
 
 class RunningMean:
@@ -68,8 +68,8 @@ def adjusted_rand_index(true_mask, pred_mask):
         raise ValueError(
         "adjusted_rand_index requires n_groups < n_points. We don't handle the special cases that can occur when you have one cluster per datapoint.")
 
-    true_group_ids = torch.argmax(true_mask, -1)
-    pred_group_ids = torch.argmax(pred_mask, -1)
+    true_group_ids = torch.argmax(true_mask, -1).int()
+    pred_group_ids = torch.argmax(pred_mask, -1).int()
     true_mask_oh = true_mask.float()
     pred_mask_oh = F.one_hot(pred_group_ids, n_pred_groups).float()
